br/gc: reorganize package structure
