+++ br/pkg/gc/manager_global.go
@@ -57,17 +57,3 @@ func (m *globalManager) DeleteServiceSafePoint(ctx context.Context, sp BRService
 	_, err := m.pdClient.UpdateServiceGCSafePoint(ctx, sp.ID, 0, 0)
 	return errors.Trace(err)
 }
-
-// UpdateServiceSafePointGlobal updates the service safe point using globalManager.
-// NOTE: This does NOT support keyspace. Use SetServiceSafePoint with storage for keyspace support.
-func UpdateServiceSafePointGlobal(ctx context.Context, pdClient pd.Client, sp BRServiceSafePoint) error {
-	mgr := newGlobalManager(pdClient)
-	return mgr.SetServiceSafePoint(ctx, sp)
-}
-
-// StartServiceSafePointKeeperGlobal starts a keeper using globalManager.
-// NOTE: This does NOT support keyspace. Use StartServiceSafePointKeeper with storage for keyspace support.
-func StartServiceSafePointKeeperGlobal(ctx context.Context, pdClient pd.Client, sp BRServiceSafePoint) error {
-	mgr := newGlobalManager(pdClient)
-	return StartServiceSafePointKeeper(ctx, sp, mgr)
-}
